refactor: windows/synch -> tools/synch
