+++ docs/tasks/2021y-02m-23d-0015-synch.md
@@ -1,0 +1,22 @@
+
+| дата начала         |   дата конца        | длительность   | исполнитель  |
+|:-------------------:|:-------------------:|:--------------:|:------------:|
+| 2020y-02m-23d 17:40 | 2020y-02m-23d 18:00 | x hours xx min | Kartonagnick |
+
+synch
+-----
+
+В версии 0.0.3  
+Класс synch является специфичным для windows  
+В его заголовочном файле присутствует зависимость от windows.h  
+Это делает класс не кроссплатформенным.  
+
+Задача
+------
+Необходимо устранить зависимость от windows.h из заголовочного файла synch.hpp так,  
+Что бы, при необходимости, можно было легко добавить поддержку ещё одной платформы.  
+
+При этом, для старых компиляторов используем собственную реализацию,
+а для компиляторов msvc2010 или старше, используем std::mutex  
+
+
