+++ test/test-600-others/sources/tools/copies/test-copies-cpp98.cpp
@@ -13,7 +13,7 @@
 #include <tools/copies.hpp>
 #include <vector>
 
-#include <tools/platform/windows/synch.hpp>
+#include <tools/synch.hpp>
 #include <process.h>
 
 namespace me = ::tools;
